chore(parametrize): centralize parametrize decorator creation and add tests
